fix: バージョン情報などいくつかの精査
